moved cardSearchQuery to context
