+++ frontend/src/context/cardSearchQuery.ts
